+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 * `orchard.inspect`: offer new `previous-sibling` and `next-sibling` functions.
   * These help navigate sequential collections one item at a time, in a streamlined fashion.
 
+### Changes
+
+* `orchard.inspect`: don't render keyword/symbol/number values as strings.
+
 ## 0.16.1 (2023-10-05)
 
 * Make some internals safer.